+++ src/Common/Hooks/UseIsTextTruncated/UseIsTextTruncated.ts
