+++ src/Common/Hooks/UseIsTextTruncated/constants.ts
@@ -1,0 +1,1 @@
+export const SUB_PIXEL_ERROR = 1
