+++ src/Common/Hooks/UseIsTextTruncated/index.ts
@@ -1,0 +1,1 @@
+export { default as useIsTextTruncated } from './UseIsTextTruncated'
