+++ src/Common/Hooks/index.ts
@@ -16,6 +16,7 @@
 
 export { useClickOutside } from './UseClickOutside/UseClickOutside'
 export { useGetUserRoles } from './UseGetUserRoles'
+export { useIsTextTruncated } from './UseIsTextTruncated'
 export * from './UseRegisterShortcut'
 export * from './useStateFilters'
 export * from './useUrlFilters'
