+++ src/Common/Tooltip/constants.tsx
@@ -17,5 +17,3 @@
 export const TOOLTIP_CONTENTS = {
     INVALID_INPUT: 'Valid input is required for all mandatory fields.',
 }
-
-export const SUB_PIXEL_ERROR = 1
