+++ src/Common/Tooltip/index.ts
@@ -17,4 +17,3 @@
 export { TOOLTIP_CONTENTS } from './constants'
 export { default as Tooltip } from './Tooltip'
 export type { TooltipProps } from './types'
-export { default as useIsTextTruncated } from './UseIsTextTruncated'
